+++ src/main/java/org/dmdev/natalliavasilyeva/api/dto/responsedto/CarUserResponseDto.java
@@ -1,5 +1,7 @@
 package org.dmdev.natalliavasilyeva.api.dto.responsedto;
 
+import java.math.BigDecimal;
+
 public class CarUserResponseDto {
 
     private String brand;
@@ -7,11 +9,10 @@ public class CarUserResponseDto {
     private String color;
     private String yearOfProduction;
     private String image;
-    private int pricePerDay;
+    private BigDecimal pricePerDay;
 
 
-
-    public CarUserResponseDto(String brand, ModelResponseDTO model, String color, String yearOfProduction, String image, int pricePerDay) {
+    public CarUserResponseDto(String brand, ModelResponseDTO model, String color, String yearOfProduction, String image, BigDecimal pricePerDay) {
         this.brand = brand;
         this.model = model;
         this.color = color;
