+++ src/main/java/org/dmdev/natalliavasilyeva/api/dto/responsedto/CategoryResponseDto.java
@@ -1,12 +1,14 @@
 package org.dmdev.natalliavasilyeva.api.dto.responsedto;
 
+import java.math.BigDecimal;
+
 public class CategoryResponseDto {
 
     private String name;
 
-    private int pricePerDay;
+    private BigDecimal pricePerDay;
 
-    public CategoryResponseDto(String name, int pricePerDay) {
+    public CategoryResponseDto(String name, BigDecimal pricePerDay) {
         this.name = name;
         this.pricePerDay = pricePerDay;
     }
