+++ src/main/java/org/dmdev/natalliavasilyeva/api/dto/responsedto/UserResponseDto.java
@@ -7,7 +7,7 @@ public class UserResponseDto extends UserShotResponseDto {
 
     private List<OrderResponseDto> orders;
 
-    public UserResponseDto(int id, String email, String name, String surname, String address, String phone, LocalDate birthday, DriverLicenseDto driverLicenseDto, List<OrderResponseDto> orders) {
+    public UserResponseDto(long id, String email, String name, String surname, String address, String phone, LocalDate birthday, DriverLicenseDto driverLicenseDto, List<OrderResponseDto> orders) {
         super(id, email, name, surname, address, phone, birthday, driverLicenseDto);
         this.orders = orders;
     }
