+++ src/main/java/org/dmdev/natalliavasilyeva/persistence/jpa/Brand.java
@@ -3,19 +3,19 @@ package org.dmdev.natalliavasilyeva.persistence.jpa;
 import java.io.Serializable;
 
 public class Brand implements Entity, Serializable {
-    private int id;
+    private long id;
     private String name;
 
     public Brand() {
     }
 
-    private Brand(int id, String name) {
+    private Brand(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
     @Override
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -25,10 +25,10 @@ public class Brand implements Entity, Serializable {
 
     public static final class Builder {
 
-        private int id;
+        private long id;
         private String name;
 
-        public Builder id(int id) {
+        public Builder id(long id) {
             this.id = id;
             return this;
         }
